+++ cellex/utils/mapping/mouse_symbol_to_mouse_ens.py
@@ -2,37 +2,45 @@ import numpy as np
 import pandas as pd
 import pkg_resources
 
-def mgi_mouse_to_ens_mouse(df_unmapped: pd.DataFrame, drop_unmapped: bool=False, verbose: bool=False) -> None:
-    """
-    Maps mouse mgi gene id's to mouse ensembl gene id's inplace.
+def mouse_symbol_to_mouse_ens(df_unmapped: pd.DataFrame, drop_unmapped: bool=False, verbose: bool=False) -> None:
+    """Map dataframe index mouse gene symbol to mouse ensemble id inplace
 
-    Args:
-        df_unmapped:    a dataframe in tidy-format.
-        drop_unmapped:  True: remove unmapped genes (rows) from df, False: keep original index
-        verbose:        explicitly print status or not
+    Parameters
+    ----------
+    df_unmapped : DataFrame
+        DataFrame with gene names of appropriate format (e.g. ensembl) as index
+    
+    drop_unmapped : bool, optional (default: False)
+        Remove unmapped genes (rows) from df, False: keep original index.
+    
+    verbose : bool, optional (default: False)
+        Print progress report.
 
-    Returns:
+    Returns
+    -------
         None
     
-    Todo:
-        * modify drop_unmapped to unmapped: {"drop", "keep", "na"}
+    TODO
+    ----
         * make one mapping-function for all cases
+        * modify drop_unmapped to unmapped: {"drop", "keep", "na"}
         * support for custom mapping file
-        * handle case for empty df
+        * handle lower/upper/mixed casee letters in index
+    
     """
 
     assert (len(df_unmapped) > 0), "Empty dataframe."
    
     if verbose:
-        print("Mapping: mouse mgi gene id's --> mouse ensembl gene id's ...")
+        print("Mapping: mouse gene symbols --> mouse ensembl gene id's ...")
+    
     resource_package = __name__
-    resource_path = 'maps/Mus_musculus.GRCm38.90.gene_name_version2ensembl.txt.gz'  # Do not use os.path.join()
+    resource_path = 'maps/Mus_musculus.GRCm38.ens_v90.gene_name_version2ensembl.txt.gz'  # Do not use os.path.join()
     resource_stream = pkg_resources.resource_stream(resource_package, resource_path)    
     df_map = pd.read_csv(resource_stream, compression='gzip', delim_whitespace=True)
-    # create dictionary for mapping mouse ensemble gene id's to human ensembl gene id's
+    # create dictionary for mapping
     map_dict = dict(zip(df_map["gene_name_optimal"].ravel(), \
                             df_map["ensembl_gene_id"].ravel()))
-    
     
     # map genes in-place,
     # i.e. indexes are replaced directly in df