+++ CHANGELOG.md
@@ -5,9 +5,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+* `utils.mapping.human_symbol_to_human_ens()`
 ### Changed
 * anova no longer uses multiprocessing. This should decrease memory usage significantly.
 * es_mu and es_sd now have index column name: "gene".
+* mapping module naming and docs
 ### Deprecated
 ### Removed
 ### Fixed